for in
we use this loop in objects..
const any name + "in" + object name;
for showing the data:
objectName[a name]
with this syntax js knows we want key in object not the index in each object .
in this loop we can't work with index.
its very similar with for of.
but we use for of for arrays and for in is for objects.
in this senario firs we loop in array and we return an object then we loop in that array with for in